refactor: Split schema between SBOR and scrypto
