+++ sbor/Cargo.toml
@@ -8,12 +8,14 @@ sbor-derive = { path = "../sbor-derive" }
 hashbrown = { version = "0.12.1", optional = true }
 serde = { version = "1.0.137", default-features = false, optional = true, features=["derive"] }
 indexmap = { version = "1.9.2", default-features = false, optional = true }
+const-sha1 = { version = "0.2.0", default-features = false, optional = true } # Chosen because of its small size and 0 transitive dependencies
 
 [features]
 # You should enable either `std` or `alloc`
 default = ["std"]
 std = ["serde?/std"]
 alloc = ["hashbrown", "serde?/alloc"]
+schema = ["indexmap", "const-sha1"]
 
 # Enable serde derives for SBOR value and type models
 serde = ["dep:serde"]
