+++ sbor/src/codec/result.rs
@@ -51,3 +51,32 @@ impl<X: CustomTypeId, D: Decoder<X>, T: Decode<X, D>, E: Decode<X, D>> Decode<X,
         }
     }
 }
+
+#[cfg(feature = "schema")]
+impl<C: CustomTypeSchema, T: Schema<C>, E: Schema<C>> Schema<C> for Result<T, E> {
+    const SCHEMA_TYPE_REF: GlobalTypeRef =
+        GlobalTypeRef::complex("Result", &[T::SCHEMA_TYPE_REF, E::SCHEMA_TYPE_REF]);
+
+    fn get_local_type_data() -> Option<LocalTypeData<C, GlobalTypeRef>> {
+        Some(LocalTypeData {
+            schema: TypeSchema::Enum {
+                variants: crate::rust::collections::indexmap::indexmap![
+                    "Ok".to_owned() => GlobalTypeRef::complex("Ok", &[T::SCHEMA_TYPE_REF]),
+                    "Err".to_owned() => GlobalTypeRef::complex("Err", &[E::SCHEMA_TYPE_REF]),
+                ],
+            },
+            naming: TypeNaming::named("Set"),
+        })
+    }
+
+    fn add_all_dependencies(aggregator: &mut SchemaAggregator<C>) {
+        aggregator.add_child_type(GlobalTypeRef::complex("Ok", &[T::SCHEMA_TYPE_REF]), || {
+            Some(LocalTypeData::named_tuple("Ok", vec![T::SCHEMA_TYPE_REF]))
+        });
+        aggregator.add_child_type(GlobalTypeRef::complex("Err", &[E::SCHEMA_TYPE_REF]), || {
+            Some(LocalTypeData::named_tuple("Err", vec![E::SCHEMA_TYPE_REF]))
+        });
+        aggregator.add_child_type_and_descendents::<T>();
+        aggregator.add_child_type_and_descendents::<E>();
+    }
+}
