+++ sbor/src/codec/unit.rs
@@ -27,3 +27,6 @@ impl<X: CustomTypeId, D: Decoder<X>> Decode<X, D> for () {
         }
     }
 }
+
+#[cfg(feature = "schema")]
+well_known_basic_schema!((), UNIT_INDEX);
