--- scrypto-abi/src/basic_impls/boolean.rs
@@ -1,3 +1,0 @@
-use crate::v2::*;
-
-well_known_schema!(bool, BOOL_INDEX);
