--- scrypto-abi/src/basic_impls/integer.rs
@@ -1,15 +1,0 @@
-use crate::v2::*;
-
-well_known_schema!(u8, U8_INDEX);
-well_known_schema!(u16, U16_INDEX);
-well_known_schema!(u32, U32_INDEX);
-well_known_schema!(u64, U64_INDEX);
-well_known_schema!(u128, U128_INDEX);
-well_known_schema!(i8, I8_INDEX);
-well_known_schema!(i16, I16_INDEX);
-well_known_schema!(i32, I32_INDEX);
-well_known_schema!(i64, I64_INDEX);
-well_known_schema!(i128, I128_INDEX);
-
-well_known_schema!(usize, U64_INDEX);
-well_known_schema!(isize, I64_INDEX);
