--- scrypto-abi/src/basic_impls/string.rs
@@ -1,4 +1,0 @@
-use crate::v2::*;
-
-well_known_schema!(String, STRING_INDEX);
-well_known_schema!(str, STRING_INDEX);
