+++ scrypto-abi/src/lib.rs
@@ -11,16 +11,4 @@ mod schema_type;
 pub use blueprint_abi::*;
 pub use schema_type::*;
 
-mod basic_impls;
-mod schema;
-mod schema_aggregator;
-mod type_ref;
-mod type_schema;
-
-pub mod v2 {
-    use super::*;
-    pub use schema::*;
-    pub use schema_aggregator::*;
-    pub use type_ref::*;
-    pub use type_schema::*;
-}
+pub mod scrypto_schema {}
